Update Migration & Models

+++ database/migrations/2023_08_22_145351_create_produks_table.php
@@ -16,9 +16,9 @@ return new class extends Migration
             $table->string('code')->unique();
             $table->string('grade_id')->nullable();
             $table->string('name', 100);
-            $table->integer('price');
-            $table->integer('estimated_sales');
-            $table->integer('stock');
+            $table->bigInteger('price');
+            $table->bigInteger('estimated_sales');
+            $table->bigInteger('stock');
             $table->string('description')->nullable();
             $table->boolean('is_active')->default(true);
             $table->timestamps();
