+++ database/migrations/2023_08_22_151824_create_production_details_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->foreignId('production_batch_id')->constrained('production_batches')->cascadeOnDelete()->cascadeOnUpdate();
             $table->foreignId('raw_material_id')->constrained('raw_materials')->cascadeOnDelete()->cascadeOnUpdate();
-            $table->integer('quantity_used');
-            $table->integer('estimated_cost');
+            $table->bigInteger('quantity_used');
+            $table->bigInteger('estimated_cost');
             $table->timestamps();
         });
     }
